requirements.txt
torch>=1.12.0
torchvision>=0.13.0
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.21.0
PyYAML>=5.4.0
tqdm>=4.62.0
scipy>=1.7.0
Pillow>=8.3.0
matplotlib>=3.4.0
torchreid>=0.2.5
filterpy>=1.4.5
streamlit